web_dev/requirements.txt
Django==4.2.19 
psycopg2-binary>=2.9,<3.0
gunicorn
django-allauth
social-auth-app-django
psycopg2-binary

